rename package name correctly
