+++ comp/trace/etwtracer/component.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023-present Datadog, Inc.
 
-// Package apmetwtracer provides ETW events to the .Net tracer
-package apmetwtracer
+// Package etwtracer provides ETW events to the .Net tracer
+package etwtracer
 
 // team: windows-agent
